ignore failed batch calls
